+++ resources/lib/WonderfulSubsBrowser.py
@@ -309,10 +309,6 @@ class WonderfulSubsBrowser(BrowserBase):
         url = self._to_url("api/media/random")
         return self._process_anime_view(url, data, "random/%d", page)
 
-    def get_anichart_airing(self, page=1):
-        from WatchlistFlavor import AniList
-        return AniList.AniChart().get_airing(page)
-
     def get_anime_metadata(self, anime_url, is_dubbed):
         info = self._get_anime_info(anime_url, is_dubbed)
         return (info["name"], info["image"])
